knn functional

+++ software/iob_knn.h
@@ -5,4 +5,4 @@ void knn_reset();
 void knn_start();	
 void knn_stop();
 void knn_init(int base_address);	
-unsigned long long distance (int x1, int x2, int y1, int y2);
+unsigned int distance (short x1, short x2, short y1, short y2);
